requirements.txt
fastapi
uvicorn
sqlalchemy
psycopg2-binary
python-multipart
python-jose[cryptography]
passlib[bcrypt]
pydantic[email]
pydantic-settings
python-dotenv
fastapi-mail

